+++ Source/Scenes/LoadingScene.hpp
@@ -26,7 +26,7 @@ private:
 	std::string_view m_loading_text; // loading text from a scene to appear bellow loading spinner to inform user of whats happening
 
 	inline static constexpr const f32 SPINNER_RADIUS = 66.66f;
-	inline static constexpr const i32 SPINNER_THICKNESS = 4;
+	inline static constexpr const f32 SPINNER_THICKNESS = 4.0f;
 };
 NS_ENIGMA_END
